added some slyses and refactored
