+++ src/components/app/app.js
@@ -1,5 +1,6 @@
 import React, {PureComponent} from 'react'
-import Restaurants from './restaurants'
+import Restaurants from '../restaurants/restaurants'
+import './app.css'
 
 class App extends PureComponent {
   render() {
