+++ src/components/dish/dish.css
@@ -1,0 +1,5 @@
+.dish {
+  margin: 20px;
+  padding: 20px;
+  border: 1px solid grey;
+}
