+++ src/components/restaurants-navigation/restaurants-navigation.css
@@ -1,0 +1,7 @@
+.ant-layout-header {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+}
