+++ src/components/restaurants-navigation/restaurants-navigation.js
@@ -1,17 +1,20 @@
 import React from 'react'
+import {Button} from 'antd'
+import './restaurants-navigation.css'
 
 function RestaurantsNavigation(props) {
   return (
-    <div>
+    <div className="header-buttons">
       {props.restaurants.map(restaurant => (
-        <button
+        <Button
           key={restaurant.id}
           onClick={() => {
             props.onRestaurantChange(restaurant.id)
           }}
+          type="default"
         >
           {restaurant.name}
-        </button>
+        </Button>
       ))}
     </div>
   )
