+++ src/components/review/review.css
@@ -1,0 +1,5 @@
+.review {
+  margin: 20px;
+  padding: 20px;
+  border: 1px solid grey;
+}
