+++ src/components/reviews/reviews.css
@@ -1,0 +1,13 @@
+.reviews-wrapper {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  border: 2px solid grey;
+  border-radius: 5px;
+  margin: 5px;
+  padding: 10px;
+}
+.reviews {
+  width: 50%;
+}
